Added sound effects and game music
Theres also menu/shop music and various sound effects in there that havent been implemented yet
