--- Assets/BoardClass.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7f6b3af84b5bcf04ca8ca84c1c9366fe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
