--- Assets/BoardControl.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d85ee361f1d67fd4da6d2512b469b57d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
