--- Assets/BulletController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1963da86d49c6c444bba87c4eb161150
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
