+++ Assets/Sound Effects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e9a4ee044d6788e4c9f381cf92b9c9bd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
